+++ flutter/example/lib/cactus_service.dart
@@ -18,7 +18,7 @@ class CactusService {
     try {
       model = await CactusVLM.init(
         modelUrl: 'https://huggingface.co/Cactus-Compute/SmolVLM2-500m-Instruct-GGUF/resolve/main/SmolVLM2-500M-Video-Instruct-Q8_0.gguf',
-        visionUrl: 'https://huggingface.co/Cactus-Compute/SmolVLM2-500m-Instruct-GGUF/resolve/main/mmproj-SmolVLM2-500M-Video-Instruct-Q8_0.gguf',
+        mmprojUrl: 'https://huggingface.co/Cactus-Compute/SmolVLM2-500m-Instruct-GGUF/resolve/main/mmproj-SmolVLM2-500M-Video-Instruct-Q8_0.gguf',
         // cactusToken: 'contact founders@cactuscompute.com for enterprise token',
         onProgress: (progress, statusText, isError) {
           status.value = statusText;